camera: fix crash on closing preview
The client_for_ion has been initialized each time the control0,
config0 or frame0 has been opened and released when closing one
of them.
This made that the ion allocations has been made using different
ion client than when freeing them which resulted in kernel oops.
This patch creates client_for_ion once when loading module and
frees it when unloading the module.
This increases the kernel memory usage by keeping the ion client
instance for the module lifetime instead of allocating it only
when it will be needed but at least the kernel crash is being fixed.